backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
torch==2.1.2
torchvision==0.16.2
opencv-python==4.9.0.80
numpy==1.26.4
pillow==10.2.0
matplotlib==3.8.4